Use npm ci when package-lock.json exists
Use `npm ci` instead of `npm install` when a `package-lock.json` file exists. Both for performance reasons and to run the exact version of `eslint` as specified in the repo.

It also adds `--ignore-scripts` to both `npm` calls to bring it in line with `yarn` and it adds `--no-package-lock` to `npm install` to keep it from creating a `package-lock.json` which wasn't there before.